+++ st2common/tests/unit/test_util_jinja.py
@@ -59,3 +59,40 @@ class JinjaUtilsRenderTestCase(unittest2.TestCase):
             allow_undefined=True)
 
         self.assertEqual(actual, expected)
+
+    def test_convert_str_to_raw(self):
+        jinja_expr = '{{foobar}}'
+        expected_raw_block = '{% raw %}{{foobar}}{% endraw %}'
+        self.assertEqual(expected_raw_block, jinja_utils.convert_jinja_to_raw_block(jinja_expr))
+
+    def test_convert_list_to_raw(self):
+        jinja_expr = ['foobar', '{{foo}}', '{{bar}}', {'foobar': '{{foobar}}'}]
+
+        expected_raw_block = [
+            'foobar',
+            '{% raw %}{{foo}}{% endraw %}',
+            '{% raw %}{{bar}}{% endraw %}',
+            {'foobar': '{% raw %}{{foobar}}{% endraw %}'}
+        ]
+
+        self.assertListEqual(expected_raw_block, jinja_utils.convert_jinja_to_raw_block(jinja_expr))
+
+    def test_convert_dict_to_raw(self):
+        jinja_expr = {
+            'var1': 'foobar',
+            'var2': ['{{foo}}', '{{bar}}'],
+            'var3': {'foobar': '{{foobar}}'}
+        }
+
+        expected_raw_block = {
+            'var1': 'foobar',
+            'var2': [
+                '{% raw %}{{foo}}{% endraw %}',
+                '{% raw %}{{bar}}{% endraw %}'
+            ],
+            'var3': {
+                'foobar': '{% raw %}{{foobar}}{% endraw %}'
+            }
+        }
+
+        self.assertDictEqual(expected_raw_block, jinja_utils.convert_jinja_to_raw_block(jinja_expr))