+++ CHANGELOG.rst
@@ -46,6 +46,7 @@ Fixed
   Reported by sibirajal.
 * Add a check to make sure action exists in the POST of the action execution API. (bug fix)
 * Fix api key generation, to use system user, when auth is disabled. (bug fix) #3578 #3593
+* Fix invocation of Mistral workflow from Action Chain with jinja in params. (bug fix) #3440
 
 2.3.1 - July 07, 2017
 ---------------------